manage left_out_region add classes to init
