+++ src/antakia_core/explanation/__init__.py
@@ -1,3 +1,6 @@
 """
 This is antakia.explanation package documentation.
 """
+
+from .explanation_method import ExplanationMethod
+from .explanations import SHAPExplanation, LIMExplanation, compute_explanations
