+++ src/antakia_core/utils/utils.py
@@ -108,7 +108,7 @@ def get_mask_comparison_color(rules_mask, selection_mask):
         'matched': 'blue',
         'error type 1': 'orange',
         'error type 2': 'red',
-        'other data': 'grey'
+        'other data': BASE_COLOR
     }
     color = pd.Series(index=selection_mask.index, dtype=str)
     color[selection_mask & rules_mask] = colors_info['matched']
@@ -120,6 +120,7 @@ def get_mask_comparison_color(rules_mask, selection_mask):
 
 # First color can't be blue, reserved for the rules - grey is reserved to background
 colors = ["red", "blue", "green", "yellow", "orange", "pink", "brown", "cyan", "black"]
+BASE_COLOR = 'grey'
 
 
 class MetaEnum(EnumMeta):
